# http/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Centrifuge OS Node API",
        "title": "Centrifuge OS Node API",
        "contact": {
            "name": "Centrifuge",
            "url": "https://github.com/centrifuge/pod",
            "email": "hello@centrifuge.io"
        },
        "license": {
            "name": "MIT"
        },
        "version": "3.0.0"
    },
    "host": "localhost:8082",
    "basePath": "/",
    "paths": {
...
        "/v2/documents": {
            "post": {
                "description": "Creates a new document.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Documents"
                ],
                "summary": "Creates a new document.",
                "operationId": "create_document_v2",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Hex encoded centrifuge ID of the account for the intended API action",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Document Create request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v2.CreateDocumentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/coreapi.DocumentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputils.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httputils.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputils.HTTPError"
                        }
                    }
                }
            }
        },
        "/v2/documents/{document_id}": {
            "patch": {
                "description": "Updates a pending document.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Documents"
                ],
                "summary": "Updates a pending document.",
                "operationId": "update_document_v2",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Hex encoded centrifuge ID of the account for the intended API action",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Document Update request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v2.UpdateDocumentRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Document Identifier",
                        "name": "document_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/coreapi.DocumentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputils.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httputils.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputils.HTTPError"
                        }
                    }
                }
            }
        },
...
        "/v2/documents/{document_id}/clone": {
            "post": {
                "description": "Creates a new cloned document from an existing Template document.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Documents"
                ],
                "summary": "Creates a new cloned document from an existing Template document.",
                "operationId": "clone_document_v2",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Hex encoded centrifuge ID of the account for the intended API action",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Document Clone request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v2.CloneDocumentRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Document Identifier",
                        "name": "document_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/coreapi.DocumentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputils.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httputils.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputils.HTTPError"
                        }
                    }
                }
            }
        },
        "/v2/documents/{document_id}/collaborators": {
            "delete": {
                "description": "Removes the collaborators from the document.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Documents"
                ],
                "summary": "Removes the collaborators from the document.",
                "operationId": "remove_collaborators",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Hex encoded centrifuge ID of the account for the intended API action",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Remove Collaborators request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v2.RemoveCollaboratorsRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Document Identifier",
                        "name": "document_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/coreapi.DocumentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputils.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httputils.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputils.HTTPError"
                        }
                    }
                }
            }
        },
        "/v2/documents/{document_id}/commit": {
            "post": {
                "description": "Commits a pending document.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Documents"
                ],
                "summary": "Commits a pending document.",
                "operationId": "commit_document_v2",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Hex encoded centrifuge ID of the account for the intended API action",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Document Identifier",
                        "name": "document_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/coreapi.DocumentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputils.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httputils.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputils.HTTPError"
                        }
                    }
                }
            }
        },
        "/v2/documents/{document_id}/committed": {
            "get": {
                "description": "Returns the latest committed document associated with docID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Documents"
                ],
                "summary": "Returns the latest committed document associated with docID.",
                "operationId": "get_committed_document",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Hex encoded centrifuge ID of the account for the intended API action",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Document Identifier",
                        "name": "document_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/coreapi.DocumentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputils.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httputils.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httputils.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputils.HTTPError"
                        }
                    }
                }
            }
        },
        "/v2/documents/{document_id}/pending": {
            "get": {
                "description": "Returns the pending document associated with docID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Documents"
                ],
                "summary": "Returns the pending document associated with docID.",
                "operationId": "get_pending_document",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Hex encoded centrifuge ID of the account for the intended API action",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Document Identifier",
                        "name": "document_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/coreapi.DocumentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputils.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httputils.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httputils.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputils.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "coreapi.AttributeMapRequest": {
            "type": "object",
            "additionalProperties": {
                "$ref": "#/definitions/coreapi.AttributeRequest"
            }
        },
...
        "coreapi.DocumentResponse": {
            "type": "object",
            "properties": {
                "attributes": {
                    "$ref": "#/definitions/coreapi.AttributeMapResponse"
                },
                "data": {
                    "type": "object"
                },
                "header": {
                    "$ref": "#/definitions/coreapi.ResponseHeader"
                },
                "scheme": {
                    "type": "string",
                    "enum": [
                        "generic",
                        "entity"
                    ]
                }
            }
        },
...
        "httputils.HTTPError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "v2.CreateDocumentRequest": {
            "type": "object",
            "properties": {
                "attributes": {
                    "$ref": "#/definitions/coreapi.AttributeMapRequest"
                },
                "data": {
                    "type": "object"
                },
                "document_id": {
                    "description": "if provided, creates the next version of the document.",
                    "type": "string"
                },
                "read_access": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "scheme": {
                    "type": "string",
                    "enum": [
                        "generic",
                        "entity"
                    ]
                },
                "write_access": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
    }
}